Updated instructions to reflect Framework and UI changes
Replaced ASP.NET Version 4.7 reference to 4.8 as 4.7 is not supported anymore. Also removed reference to "Allow other resources to access this server" as this is not being configured at this time/location anymore. Added instructions to configure this later. Will replace files to reflect these changes in screenshots as well.